Configure /logout
Logs out a user via POST /logout. As Basic Auth is being used, the
browser has to be made forget about the credentials (sent with
every request) to make this work.

An example login is started with the auth proxy.

Serves requests now at /api/** to avoid mixing non-forwarding calls